agent/rag_demo/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
openai
langchain-text-splitters
fastapi
uvicorn
python-multipart
rank-bm25
